+++ assembly/src/ast/mod.rs
@@ -4,28 +4,22 @@ use super::{
     Serializable, SliceReader, SourceLocation, StarkField, String, ToString, Token, TokenStream,
     Vec, MAX_LABEL_LEN,
 };
-use core::{fmt::Display, iter, ops::RangeBounds, str::from_utf8};
+use core::{iter, str::from_utf8};
+use vm_core::utils::bound_into_included_u64;
 
-mod body;
-use body::CodeBody;
 mod nodes;
-use crate::utils::bound_into_included_u64;
-pub use nodes::display_hex_bytes;
 pub use nodes::{AdviceInjector, Instruction, Node};
-mod context;
-use context::ParserContext;
-mod labels;
-use labels::CONSTANT_LABEL_PARSER;
-pub use labels::{decode_hex_rpo_digest_label, NAMESPACE_LABEL_PARSER, PROCEDURE_LABEL_PARSER};
+
+mod code_body;
+pub use code_body::CodeBody;
+
 mod invocation_target;
 pub use invocation_target::InvocationTarget;
 
-mod adv_ops;
-mod field_ops;
-mod io_ops;
-mod serde;
-mod stack_ops;
-mod u32_ops;
+mod parsers;
+use parsers::{parse_constants, parse_imports, ParserContext};
+
+pub use parsers::{NAMESPACE_LABEL_PARSER, PROCEDURE_LABEL_PARSER};
 
 #[cfg(test)]
 pub mod tests;
@@ -200,7 +194,7 @@ impl ProgramAst {
 
     /// Loads the [SourceLocation] from the `source`.
     ///
-    /// It expects the `start` location at the first position, and will subsequentially load the
+    /// It expects the `start` location at the first position, and will subsequently load the
     /// body via [CodeBody::load_source_locations]. Finally, it will load the local procedures via
     /// [ProcedureAst::load_source_locations].
     pub fn load_source_locations<R: ByteReader>(
@@ -232,7 +226,7 @@ impl ProgramAst {
     }
 }
 
-// LIBRARY MODULE AST
+// MODULE AST
 // ================================================================================================
 
 /// An abstract syntax tree (AST) of a Miden code module.
@@ -552,84 +546,6 @@ impl Deserializable for ProcedureAst {
     }
 }
 
-// PARSERS
-// ================================================================================================
-
-/// Parses all `use` statements into a map of imports which maps a module name (e.g., "u64") to
-/// its fully-qualified path (e.g., "std::math::u64").
-fn parse_imports(tokens: &mut TokenStream) -> Result<BTreeMap<String, LibraryPath>, ParsingError> {
-    let mut imports = BTreeMap::<String, LibraryPath>::new();
-    // read tokens from the token stream until all `use` tokens are consumed
-    while let Some(token) = tokens.read() {
-        match token.parts()[0] {
-            Token::USE => {
-                let module_path = token.parse_use()?;
-                let module_name = module_path.last();
-                if imports.contains_key(module_name) {
-                    return Err(ParsingError::duplicate_module_import(token, &module_path));
-                }
-
-                imports.insert(module_name.to_string(), module_path);
-
-                // consume the `use` token
-                tokens.advance();
-            }
-            _ => break,
-        }
-    }
-
-    Ok(imports)
-}
-
-/// Parses all `const` statements into a map which maps a const name to a value
-fn parse_constants(tokens: &mut TokenStream) -> Result<LocalConstMap, ParsingError> {
-    // instantiate new constant map for this module
-    let mut constants = LocalConstMap::new();
-
-    // iterate over tokens until we find a const declaration
-    while let Some(token) = tokens.read() {
-        match token.parts()[0] {
-            Token::CONST => {
-                let (name, value) = parse_constant(token)?;
-
-                if constants.contains_key(&name) {
-                    return Err(ParsingError::duplicate_const_name(token, &name));
-                }
-
-                constants.insert(name, value);
-                tokens.advance();
-            }
-            _ => break,
-        }
-    }
-
-    Ok(constants)
-}
-
-/// Parses a constant token and returns a (constant_name, constant_value) tuple
-fn parse_constant(token: &Token) -> Result<(String, u64), ParsingError> {
-    match token.num_parts() {
-        0 => unreachable!(),
-        1 => Err(ParsingError::missing_param(token)),
-        2 => {
-            let const_declaration: Vec<&str> = token.parts()[1].split('=').collect();
-            match const_declaration.len() {
-                0 => unreachable!(),
-                1 => Err(ParsingError::missing_param(token)),
-                2 => {
-                    let name = CONSTANT_LABEL_PARSER
-                        .parse_label(const_declaration[0])
-                        .map_err(|err| ParsingError::invalid_const_name(token, err))?;
-                    let value = parse_const_value(token, const_declaration[1])?;
-                    Ok((name.to_string(), value))
-                }
-                _ => Err(ParsingError::extra_param(token)),
-            }
-        }
-        _ => Err(ParsingError::extra_param(token)),
-    }
-}
-
 // HELPER FUNCTIONS
 // ================================================================================================
 
@@ -639,95 +555,4 @@ fn sort_procs_into_vec(proc_map: LocalProcMap) -> Vec<ProcedureAst> {
     procedures.sort_by_key(|(idx, _proc)| *idx);
 
     procedures.into_iter().map(|(_idx, proc)| proc).collect()
-}
-
-/// Parses a param from the op token with the specified type and index. If the param is a constant
-/// label, it will be looked up in the provided constant map.
-fn parse_param_with_constant_lookup<R>(
-    op: &Token,
-    param_idx: usize,
-    constants: &LocalConstMap,
-) -> Result<R, ParsingError>
-where
-    R: TryFrom<u64> + core::str::FromStr,
-{
-    let param_str = op.parts()[param_idx];
-    match CONSTANT_LABEL_PARSER.parse_label(param_str) {
-        Ok(_) => {
-            let constant = constants
-                .get(param_str)
-                .cloned()
-                .ok_or_else(|| ParsingError::const_not_found(op))?;
-            constant
-                .try_into()
-                .map_err(|_| ParsingError::const_conversion_failed(op, core::any::type_name::<R>()))
-        }
-        Err(_) => parse_param::<R>(op, param_idx),
-    }
-}
-
-/// Parses a param from the op token with the specified type.
-fn parse_param<I: core::str::FromStr>(op: &Token, param_idx: usize) -> Result<I, ParsingError> {
-    let param_value = op.parts()[param_idx];
-
-    let result = match param_value.parse::<I>() {
-        Ok(i) => i,
-        Err(_) => return Err(ParsingError::invalid_param(op, param_idx)),
-    };
-
-    Ok(result)
-}
-
-/// Parses a constant value and ensures it falls within bounds specified by the caller
-fn parse_const_value(op: &Token, const_value: &str) -> Result<u64, ParsingError> {
-    let result = const_value
-        .parse::<u64>()
-        .map_err(|err| ParsingError::invalid_const_value(op, const_value, &err.to_string()))?;
-
-    let range = 0..Felt::MODULUS;
-    range.contains(&result).then_some(result).ok_or_else(|| ParsingError::invalid_const_value(op, const_value, format!(
-        "constant value must be greater than or equal to {lower_bound} and less than or equal to {upper_bound}", lower_bound = bound_into_included_u64(range.start_bound(), true),
-        upper_bound = bound_into_included_u64(range.end_bound(), false)
-    )
-    .as_str(),))
-}
-
-/// Parses a param from the op token with the specified type and ensures that it falls within the
-/// bounds specified by the caller.
-fn parse_checked_param<I, R>(op: &Token, param_idx: usize, range: R) -> Result<I, ParsingError>
-where
-    I: core::str::FromStr + Ord + Clone + Into<u64> + Display,
-    R: RangeBounds<I>,
-{
-    let param_value = op.parts()[param_idx];
-
-    let result = match param_value.parse::<I>() {
-        Ok(i) => i,
-        Err(_) => return Err(ParsingError::invalid_param(op, param_idx)),
-    };
-
-    // check that the parameter is within the specified bounds
-    range.contains(&result).then_some(result).ok_or_else(||
-        ParsingError::invalid_param_with_reason(
-            op,
-            param_idx,
-            format!(
-                "parameter value must be greater than or equal to {lower_bound} and less than or equal to {upper_bound}", lower_bound = bound_into_included_u64(range.start_bound(), true),
-                upper_bound = bound_into_included_u64(range.end_bound(), false)
-            )
-            .as_str(),
-        )
-    )
-}
-
-/// Returns an error if the passed in value is 0.
-///
-/// This is intended to be used when parsing instructions which need to perform division by
-/// immediate value.
-fn check_div_by_zero(value: u64, op: &Token, param_idx: usize) -> Result<(), ParsingError> {
-    if value == 0 {
-        Err(ParsingError::invalid_param_with_reason(op, param_idx, "division by zero"))
-    } else {
-        Ok(())
-    }
 }
