+++ assembly/src/ast/nodes/serde/mod.rs
@@ -1,9 +1,8 @@
-use super::{
-    ByteReader, ByteWriter, CodeBody, Deserializable, DeserializationError, Felt, Instruction,
-    Node, ProcedureId, RpoDigest, Serializable,
-};
+use super::{CodeBody, Felt, Instruction, Node, ProcedureId, RpoDigest};
 use crate::MAX_PUSH_INPUTS;
 use num_enum::TryFromPrimitive;
+
+use vm_core::utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
 mod deserialization;
 mod serialization;
