+++ assembly/src/lib.rs
@@ -4,6 +4,8 @@
 #[macro_use]
 extern crate alloc;
 
+use core::fmt;
+
 use vm_core::{
     code_blocks::CodeBlock,
     crypto,
@@ -21,14 +23,16 @@ mod procedures;
 use procedures::{CallSet, Procedure};
 pub use procedures::{ProcedureId, ProcedureName};
 
-mod parsers;
-use parsers::{display_hex_bytes, NAMESPACE_LABEL_PARSER, PROCEDURE_LABEL_PARSER};
-pub use parsers::{InvocationTarget, ModuleAst, ProcedureAst, ProgramAst};
+//mod parsers;
+//pub use parsers::{InvocationTarget, ModuleAst, ProcedureAst, ProgramAst};
 
-pub mod ast {
-    pub use crate::parsers::{Instruction, ModuleAst, Node, ProcedureAst, ProgramAst};
-    pub use crate::procedures::ProcedureName;
-}
+//pub mod ast2 {
+//    pub use crate::parsers::{Instruction, ModuleAst, Node, ProcedureAst, ProgramAst};
+//    pub use crate::procedures::ProcedureName;
+//}
+
+pub mod ast;
+use ast::{NAMESPACE_LABEL_PARSER, PROCEDURE_LABEL_PARSER};
 
 pub use vm_core::utils::{
     ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable, SliceReader,
@@ -77,3 +81,12 @@ const MAX_LABEL_LEN: usize = 255;
 /// The required length of the hexadecimal representation for an input value when more than one hex
 /// input is provided to `push` masm operation without period separators.
 const HEX_CHUNK_SIZE: usize = 16;
+
+/// Builds a hex string from a byte slice
+pub fn display_hex_bytes(f: &mut fmt::Formatter<'_>, bytes: &[u8]) -> fmt::Result {
+    write!(f, "0x")?;
+    for byte in bytes {
+        write!(f, "{byte:02x}")?;
+    }
+    Ok(())
+}
